// oracle_tai/oracle/types.ts
export interface Poem {
  id: number;
  content: string[];
}

export interface InterpretationResult {
  thaiPoem: string[];
  markdown: string;
}

export enum AppStep {
  QUESTION = 'question',
  PICKER = 'picker',
  RESULT = 'result',
}

export interface RandomSeed {
  letter: string;
  number: number;
}